reservaSalas/requirements.txt
flask==2.0.1
pymysql==1.0.2
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.28.0
python-dotenv==0.19.1
flask-oauthlib==0.9.6
authlib==1.0.1
requests==2.26.0
gunicorn==20.1.0 